Rewriting Thrift connection handling
The diskstorage.cassandra.thrift part of this branch is in the middle
of a refactoring effort to reduce the total number of Apache
commons-pool instances per StorageManager to one and to choose an
endpoint for Cassandra operations based on a basic backward-looking
activity measure devised by Matthias.

This commit will not compile.  Substantial reimplementation and
cleanup work remains to be done in CassandraThriftKCVS and
CassandraThriftSM.